pyutils/mask-faster-rcnn/requirements.txt
numpy
h5py
matplotlib
scipy
scikit-image
jupyter
tensorboardX
torchvision

